Polymorphic container and Abstract GeoPoint
Zone geometries is a container of smart pointers because Geometry is an abstract class, so we can't call virtual methods on the base class. We can now call "contains" on each element of the container by following the pointer, which points to a derived class.

GeoPoint is derived as IGCPoint, which will allow constructing GeoPoint from OpenAir files directly by instantiating another derived class, rather than checking the parsing method to use.
Note that GeoPoint require a default constructor because of the variable assignement of OpenAir, which might or might not appear in a record.