Add elevation requests
This allows to get ground altitudes of the flight if the airspace needs to be checked against these altitudes.
The request is only made if the airspace file requires it, and if the user has a api key to the elevation service.
If the response of the service does not match the number of points in the flight, agl zones will not be checked.